ME-637: Moved IJsonSerializable to event.py.
