+++ clare/clare/crawler/logging/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from .i_json_serializable import IJsonSerializable
+from .event import IJsonSerializable
 
 __all__ = ['IJsonSerializable']
